backend/requirements.txt
# FastAPI ve Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Veritabanı
sqlalchemy==2.0.23
asyncpg==0.29.0  # PostgreSQL için
psycopg2-binary==2.9.9  # PostgreSQL için
firebase-admin==6.2.0  # Firestore için

# RSS ve HTTP İstekleri
feedparser==6.0.10
requests==2.31.0
aiohttp==3.9.1

# AI ve Özet Çıkarma (Google Gemini)
google-generativeai==0.3.2

# E-posta Gönderimi (Resend - SendGrid alternatifi)
resend==0.8.0

# Zamanlayıcı ve Görev Yönetimi
apscheduler==3.10.4
pytz==2023.3  # Zaman dilimi desteği

# Veri İşleme ve Validasyon
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Loglama ve Monitoring
loguru==0.7.2

# HTML Parsing
beautifulsoup4==4.12.2
lxml==4.9.3 